Fix cooler count display to show whole numbers
Correctly cast PostgreSQL SUM() results to integers in the coolers summary route to prevent floating-point display issues.

+++ server/routes/coolers.ts
@@ -196,8 +196,8 @@ coolerInventoryRouter.get('/summary', async (req: AuthenticatedRequest, res: Res
       .select({
         coolerTypeId: coolerTypes.id,
         coolerTypeName: coolerTypes.name,
-        totalQuantity: sql<number>`COALESCE(SUM(${coolerInventory.quantity}), 0)`,
-        locationCount: sql<number>`COUNT(DISTINCT ${coolerInventory.hostHomeId})`,
+        totalQuantity: sql<number>`COALESCE(SUM(${coolerInventory.quantity})::integer, 0)`,
+        locationCount: sql<number>`COUNT(DISTINCT ${coolerInventory.hostHomeId})::integer`,
       })
       .from(coolerTypes)
       .leftJoin(
